+++ tests/testthat/sql/getResults.sql
@@ -1,0 +1,5 @@
+{TYPEC INT[] @cohort_id}
+{TYPEC TEXT @cohort}
+{TYPEC TEXT @result_schema}
+
+SELECT * FROM @result_schema.@cohort where cohort_definition_id IN (@cohort_id)