+++ tests/testthat/sql/getResultsTest.sql
@@ -1,0 +1,2 @@
+SELECT * FROM @results_database_schema.@cohort
+WHERE cohort_id = @cohort_id