+++ tests/testthat/test-QueryNamespace.R
@@ -61,8 +61,39 @@ test_that("test setConnectionHandler and getConnectionHandler functions", {
   result <- cohortNamespace$queryDb("SELECT * FROM @result_schema.@cohort WHERE cohort_definition_id = @cohort_id", cohort_id = 1)
   expect_true(is.data.frame(result))
 
+  cohortNamespace$addQueryFile("sql/get_results.sql", snakeCaseToCamelCaseFileNames = TRUE)
+  result <- cohortNamespace$getResults(cohort_id = 1)
+  expect_true(is.data.frame(result))
+
   result <- cohortNamespace$executeSql("SELECT * FROM @result_schema.@cohort WHERE cohort_definition_id = @cohort_id", cohort_id = 1)
   expect_type(result, "NULL")
+
+
+  qns <- QueryNamespace$new(
+    tableSpecification = tableSpecification,
+    result_schema = "main",
+    tablePrefix = "cd_",
+    queryFiles = c("sql/getResultsTest.sql", "sql/getResults.sql"),
+    connectionHandler = connectionHandler
+  )
+
+  expect_true("getResultsTest" %in% names(qns))
+
+  expect_error(qns$addQueryFile("nonExistenFile.sql"))
+  expect_error(qns$addQueryFile("sql/notAnSqlFile.txt"))
+
+  expect_error(qns$addQueryFile("sql/get_results.sql", snakeCaseToCamelCaseFileNames = TRUE))
+
+  expect_true("getResults" %in% names(qns))
+  checkmate::expect_data_frame(qns$getResults(cohort_id = c(1,2,3)))
+
+  qns$addQueryFile("sql/createTestData.sql", execFunction = TRUE)
+  expect_true("createTestData" %in% names(qns))
+
+  qns$createTestData()
+
+  checkmate::expect_data_frame(qns$queryDb("SELECT * FROM @result_schema.foo"))
+  checkmate::expect_data_frame(qns$queryDb("SELECT * FROM @result_schema.foo2"))
 })
 
 test_that("create helper function works", {